Two siblings the earlier fixes missed, both found by the local sweep
A 64-core run of tools/verify/run_all.sh, 3,479 functions checked, and two
of the 110 exported-arithmetic candidates are bugs already fixed ELSEWHERE
in the same library:

  s_remquo.c:123    lx = (hx<<(32-n))|(lx>>n);
  s_remquo.c:120    lx = (lx>>n)|((u_int32_t)hx<<(32-n));

identical to e_fmod.c:116 and e_fmod.c:113, and with the same three-line
gap between the wrong one and the right one.

  s_nexttowardf.cpp:37   SET_FLOAT_WORD(x,(uy.bits.sign<<31)|1);
  s_nexttoward.c:46      INSERT_WORDS(x,(u_int32_t)uy.bits.sign<<31,1);

the second of which I fixed two commits ago and the first of which I did
not look for. uy.bits.sign is `unsigned int :1`, which promotes to int, so
this is 1 << 31 signed.

That is the fourth and fifth time this shape has appeared with its own
correct form already sitting nearby:

  stdc_first_leading_zero.c had the cast the other three stdbit files did not
  e_fmod.c:113            had it, e_fmod.c:116 did not
  s_nexttoward.c          fixed, s_nexttowardf.cpp not looked at
  s_remquo.c:120          has it, s_remquo.c:123 did not

The lesson is not about these files. Each time I fixed the instance the
tool named and did not grep for the pattern. The sweep did.

Bit-preserving: 2,200,002 cases across every n in [21,31] and both sign
values, zero mismatches, under UBSan. Both port gates still pass -
s_nexttowardf.cpp is a landed port, so the symbol and diagnostic checks
apply to it.

Tree-wide grep for either pattern without a cast now returns nothing.

Nine UB defects fixed.

+++ hbsd/src/lib/msun/src/s_nexttowardf.cpp
@@ -34,7 +34,7 @@ nexttowardf(float x, long double y)
 	   return x+y;	/* x or y is nan */
 	if(x==y) return (float)y;		/* x=y, return y */
 	if(ix==0) {				/* x == 0 */
-	    SET_FLOAT_WORD(x,(uy.bits.sign<<31)|1);/* return +-minsubnormal */
+	    SET_FLOAT_WORD(x,((u_int32_t)uy.bits.sign<<31)|1);/* return +-minsubnormal */
 	    t = x*x;
 	    if(t==x) return t; else return x;	/* raise underflow flag */
 	}
